+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/LocalDataProvidersIntegrationTest.kt
@@ -5,6 +5,7 @@ import com.mapbox.search.record.FavoritesDataProvider
 import com.mapbox.search.record.HistoryDataProvider
 import com.mapbox.search.tests_support.BlockingOnDataProviderEngineRegisterListener
 import com.mapbox.search.tests_support.record.clearBlocking
+import okhttp3.mockwebserver.MockWebServer
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
@@ -19,14 +20,24 @@ internal class LocalDataProvidersIntegrationTest : BaseTest() {
     override fun setUp() {
         super.setUp()
 
-        MapboxSearchSdk.initializeInternal(
-            application = targetApplication,
+        val searchEngineSettings = SearchEngineSettings(
+            applicationContext = targetApplication,
             accessToken = DEFAULT_TEST_ACCESS_TOKEN,
             locationEngine = FixedPointLocationEngine(DEFAULT_TEST_USER_LOCATION),
+            singleBoxSearchBaseUrl = MockWebServer().url("").toString()
+        )
+
+        MapboxSearchSdk.initializeInternal(
+            searchEngineSettings = searchEngineSettings,
+            offlineSearchEngineSettings = OfflineSearchEngineSettings(
+                applicationContext = targetApplication,
+                accessToken = DEFAULT_TEST_ACCESS_TOKEN,
+                locationEngine = FixedPointLocationEngine(DEFAULT_TEST_USER_LOCATION),
+            ),
             allowReinitialization = true,
         )
 
-        searchEngine = MapboxSearchSdk.createSearchEngine(ApiType.SBS, SearchEngineSettings(), useSharedCoreEngine = true)
+        searchEngine = MapboxSearchSdk.createSearchEngine(ApiType.SBS, searchEngineSettings, useSharedCoreEngine = true)
 
         historyDataProvider = MapboxSearchSdk.serviceProvider.historyDataProvider()
         historyDataProvider.clearBlocking()