+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/SearchEngineDataProvidersIntegrationTest.kt
@@ -32,15 +32,20 @@ internal class SearchEngineDataProvidersIntegrationTest : BaseTest() {
         mockServer = MockWebServer()
 
         searchEngineSettings = SearchEngineSettings(
+            applicationContext = targetApplication,
+            accessToken = DEFAULT_TEST_ACCESS_TOKEN,
+            locationEngine = FixedPointLocationEngine(DEFAULT_TEST_USER_LOCATION),
             singleBoxSearchBaseUrl = mockServer.url("").toString(),
             geocodingEndpointBaseUrl = mockServer.url("").toString()
         )
 
         MapboxSearchSdk.initializeInternal(
-            application = targetApplication,
-            accessToken = DEFAULT_TEST_ACCESS_TOKEN,
-            locationEngine = FixedPointLocationEngine(DEFAULT_TEST_USER_LOCATION),
             searchEngineSettings = searchEngineSettings,
+            offlineSearchEngineSettings = OfflineSearchEngineSettings(
+                applicationContext = targetApplication,
+                accessToken = DEFAULT_TEST_ACCESS_TOKEN,
+                locationEngine = FixedPointLocationEngine(DEFAULT_TEST_USER_LOCATION),
+            ),
             allowReinitialization = true,
         )