+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchEngineImpl.kt
@@ -23,6 +23,7 @@ import com.mapbox.search.result.ServerSearchSuggestion
 import com.mapbox.search.result.mapToCore
 import java.util.concurrent.Executor
 import java.util.concurrent.ExecutorService
+import java.util.concurrent.Executors
 
 internal class SearchEngineImpl(
     override val apiType: ApiType,
@@ -31,7 +32,7 @@ internal class SearchEngineImpl(
     private val historyService: HistoryService,
     private val requestContextProvider: SearchRequestContextProvider,
     private val searchResultFactory: SearchResultFactory,
-    private val engineExecutorService: ExecutorService,
+    private val engineExecutorService: ExecutorService = DEFAULT_EXECUTOR,
     private val indexableDataProvidersRegistry: IndexableDataProvidersRegistry,
 ) : BaseSearchEngine(), SearchEngine {
 
@@ -350,5 +351,11 @@ internal class SearchEngineImpl(
             executor = executor,
             callback = callback,
         )
+    }
+
+    private companion object {
+        val DEFAULT_EXECUTOR: ExecutorService = Executors.newSingleThreadExecutor { runnable ->
+            Thread(runnable, "SearchEngine executor")
+        }
     }
 }