+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/OfflineSearchEngineSettingsTest.kt
@@ -1,10 +1,50 @@
 package com.mapbox.search
 
+import android.content.Context
+import com.mapbox.android.core.location.LocationEngine
+import com.mapbox.common.TileStore
+import com.mapbox.search.common.tests.CopyVerifier
+import com.mapbox.search.common.tests.ReflectionObjectsFactory
+import com.mapbox.search.common.tests.ToStringVerifier
+import com.mapbox.search.tests_support.MockedTypesObjectCreators
+import com.mapbox.search.tests_support.SdkCustomTypeObjectCreators
 import com.mapbox.test.dsl.TestCase
+import io.mockk.mockk
+import nl.jqno.equalsverifier.EqualsVerifier
 import org.junit.jupiter.api.TestFactory
 import java.net.URI
 
 internal class OfflineSearchEngineSettingsTest {
+
+    private val reflectionObjectFactory = ReflectionObjectsFactory(
+        extraCreators = SdkCustomTypeObjectCreators.ALL_CREATORS + MockedTypesObjectCreators.ALL_CREATORS
+    )
+
+    @TestFactory
+    fun `Test generated equals(), hashCode(), copy(), and toString() methods`() = TestCase {
+        Given("OfflineSearchEngineSettings class") {
+            When("equals(), hashCode(), copy(), and toString() called") {
+                Then("equals() and hashCode() functions should use every declared property") {
+                    EqualsVerifier.forClass(OfflineSearchEngineSettings::class.java).verify()
+                }
+
+                Then("toString() function should use every declared property") {
+                    ToStringVerifier(
+                        clazz = OfflineSearchEngineSettings::class,
+                        objectsFactory = reflectionObjectFactory,
+                        includeAllProperties = false
+                    ).verify()
+                }
+
+                Then("copy() function should use every declared property") {
+                    CopyVerifier(
+                        clazz = OfflineSearchEngineSettings::class,
+                        objectsFactory = reflectionObjectFactory,
+                    ).verify()
+                }
+            }
+        }
+    }
 
     @TestFactory
     fun `Check OfflineSearchSettings default settings`() = TestCase {
@@ -35,8 +75,17 @@ internal class OfflineSearchEngineSettingsTest {
     fun `Check OfflineSearchSettings default builder`() = TestCase {
         Given("OfflineSearchSettings builder") {
             When("Build new settings with default values") {
-                val actual = OfflineSearchEngineSettings.Builder().build()
-                val expected = OfflineSearchEngineSettings()
+                val actual = OfflineSearchEngineSettings.Builder(TEST_MOCKED_CONTEXT, TEST_ACCESS_TOKEN)
+                    .locationEngine(TEST_MOCKED_LOCATION_ENGINE)
+                    .tileStore(TEST_MOCKED_TILE_STORE)
+                    .build()
+
+                val expected = OfflineSearchEngineSettings(
+                    applicationContext = TEST_MOCKED_CONTEXT,
+                    accessToken = TEST_ACCESS_TOKEN,
+                    locationEngine = TEST_MOCKED_LOCATION_ENGINE,
+                    tileStore = TEST_MOCKED_TILE_STORE
+                )
 
                 Then("Settings should be equal", expected, actual)
             }
@@ -47,13 +96,20 @@ internal class OfflineSearchEngineSettingsTest {
     fun `Check OfflineSearchSettings builder with all values set`() = TestCase {
         Given("OfflineSearchSettings builder") {
             When("Build new settings with test values") {
-                val actual = OfflineSearchEngineSettings.Builder().run {
-                    tilesBaseUri(TEST_DEFAULT_ENDPOINT_URI)
-                    build()
-                }
+                val actual = OfflineSearchEngineSettings.Builder(TEST_MOCKED_CONTEXT, TEST_ACCESS_TOKEN)
+                    .locationEngine(TEST_MOCKED_LOCATION_ENGINE)
+                    .tileStore(TEST_MOCKED_TILE_STORE)
+                    .tilesBaseUri(TEST_DEFAULT_ENDPOINT_URI)
+                    .viewportProvider(TEST_MOCKED_VIEWPORT_PROVIDER)
+                    .build()
 
                 val expected = OfflineSearchEngineSettings(
+                    applicationContext = TEST_MOCKED_CONTEXT,
+                    accessToken = TEST_ACCESS_TOKEN,
+                    locationEngine = TEST_MOCKED_LOCATION_ENGINE,
                     tilesBaseUri = TEST_DEFAULT_ENDPOINT_URI,
+                    tileStore = TEST_MOCKED_TILE_STORE,
+                    viewportProvider = TEST_MOCKED_VIEWPORT_PROVIDER,
                 )
 
                 Then("Settings should be equal", expected, actual)
@@ -66,7 +122,12 @@ internal class OfflineSearchEngineSettingsTest {
         Given("OfflineSearchSettings builder") {
             When("Object created with toBuilder()") {
                 val settings = OfflineSearchEngineSettings(
+                    applicationContext = TEST_MOCKED_CONTEXT,
+                    accessToken = TEST_ACCESS_TOKEN,
+                    locationEngine = TEST_MOCKED_LOCATION_ENGINE,
                     tilesBaseUri = TEST_DEFAULT_ENDPOINT_URI,
+                    tileStore = TEST_MOCKED_TILE_STORE,
+                    viewportProvider = TEST_MOCKED_VIEWPORT_PROVIDER,
                 )
 
                 Then("Settings should be equal", settings, settings.toBuilder().build())
@@ -76,5 +137,10 @@ internal class OfflineSearchEngineSettingsTest {
 
     private companion object {
         val TEST_DEFAULT_ENDPOINT_URI: URI = URI.create("https://api-offline-search-staging.tilestream.net")
+        const val TEST_ACCESS_TOKEN = "test token"
+        val TEST_MOCKED_CONTEXT: Context = mockk(relaxed = true)
+        val TEST_MOCKED_LOCATION_ENGINE: LocationEngine = mockk(relaxed = true)
+        val TEST_MOCKED_TILE_STORE: TileStore = mockk(relaxed = true)
+        val TEST_MOCKED_VIEWPORT_PROVIDER: ViewportProvider = mockk(relaxed = true)
     }
 }